old-tests/submission/Infraware/Forms/contents/Forms/output_validationmessage.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_validationMessage</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <output name=o id='output_id'></output>
	</form>
	<script>

		var output = document.getElementById("output_id");

		if (typeof(output.validationMessage) == "string") {
			test(function() {
				assert_equals(output.validationMessage, "", "validationMessage attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("validationMessage attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
